Basic Login functionality

+++ backend/src/main/java/com/example/backend/repository/DepartmentRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DepartmentRepo extends JpaRepository<Department,Integer> {
+public interface DepartmentRepo extends JpaRepository<Department,Long> {
 
     // TODO: Implement repository logic (JPA/Hibernate/etc.)
     Optional<Department> findByName(String name);
