+++ backend/src/main/java/com/example/backend/repository/StudentRepresentativeRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentRepresentativeRepo extends JpaRepository<StudentRepresentative, Integer> {
+public interface StudentRepresentativeRepo extends JpaRepository<StudentRepresentative, Long> {
 
     // TODO: Implement repository logic (JPA/Hibernate/etc.)
